+++ R/jamenrich-base.r
@@ -450,6 +450,7 @@ enrichDF2enrichResult <- function
 multiEnrichMap <- function
 (enrichList,
  geneHitList=NULL,
+ geneHitIM=NULL,
  colorV=NULL,
  nrow=NULL,
  ncol=NULL,
@@ -710,9 +711,33 @@ multiEnrichMap <- function
          "geneIM <- list2im(geneHitList)");
    }
    geneIM <- list2im(geneHitList)[,,drop=FALSE];
-   if (verbose) {
-      jamba::printDebug("multiEnrichMap(): ",
-         "geneIMcolors <- matrix2heatColors(geneIM)");
+
+   #####################################################################
+   ## geneIMdirection (if geneHitIM is supplied)
+   geneIMdirection <- NULL;
+   if (length(geneHitIM) > 0 &&
+         any(rownames(geneHitIM) %in% rownames(geneIM)) &&
+         any(colnames(geneHitIM) %in% colnames(geneIM))) {
+      geneIMdirection <- (abs(geneHitIM) > 0) * 1;
+      shared_rows <- intersect(rownames(geneHitIM),
+         rownames(geneIMdirection));
+      shared_columns <- intersect(colnames(geneHitIM),
+         colnames(geneIMdirection));
+      geneIMdirection[shared_rows, shared_columns] <- geneHitIM[shared_rows, shared_columns];
+      if (!all(rownames(geneIMdirection) %in% shared_rows)) {
+         missing_rows <- setdiff(rownames(geneIMdirection), shared_rows);
+         jamba::printDebug("multiEnrichMap(): ",
+            "geneHitIM does not contain ",
+            jamba::formatInt(length(missing_rows)),
+            " rows present in geneIM, default values will use 1.");
+      }
+      if (!all(colnames(geneIMdirection) %in% shared_columns)) {
+         missing_columns <- setdiff(colnames(geneIMdirection), shared_columns);
+         jamba::printDebug("multiEnrichMap(): ",
+            "geneHitIM does not contain ",
+            jamba::formatInt(length(missing_columns)),
+            " columns present in geneIM, default values will use 1.");
+      }
    }
 
    #####################################################################
@@ -777,7 +802,10 @@ multiEnrichMap <- function
          jamba::printDebug("setdiff(newGenes, rownames(geneIM)):", setdiff(newGenes, rownames(geneIM)));
          jamba::printDebug("setdiff(newGenes, origGenes):", setdiff(newGenes, origGenes));
       }
-      geneIM <- geneIM[rownames(geneIM) %in% newGenes,,drop=FALSE];
+      geneIM <- subset(geneIM, rownames(geneIM) %in% newGenes);
+      if (length(geneIMdirection) > 0) {
+         geneIMdirection <- subset(geneIMdirection, rownames(geneIMdirection) %in% newGenes);
+      }
       if (verbose) {
          jamba::printDebug("multiEnrichMap(): ",
             "nrow(geneIM) after:",
@@ -798,15 +826,32 @@ multiEnrichMap <- function
 
    #####################################################################
    ## geneIM colors
-   geneIMcolors <- colorjam::matrix2heatColors(x=geneIM,
-      transformFunc=c,
-      colorV=colorV,
-      shareNumLimit=FALSE,
-      numLimit=1,
-      trimRamp=c(4,3));
+   if (verbose) {
+      jamba::printDebug("multiEnrichMap(): ",
+         "creating geneIMcolors");
+   }
+   # change this code to use actual colorV colors, no gradient necessary
+   geneIMcolors <- do.call(cbind, lapply(jamba::nameVector(colnames(geneIM)), function(icolname){
+      ifelse(geneIM[,icolname] == 0,
+         "#FFFFFF",
+         rep(colorV[icolname], nrow(geneIM)))
+   }))
+   # just to be sure, set rownames and colnames to match geneIM
+   rownames(geneIMcolors) <- rownames(geneIM);
+   colnames(geneIMcolors) <- colnames(geneIM);
+   # geneIMcolors <- colorjam::matrix2heatColors(x=geneIM,
+   #    transformFunc=c,
+   #    colorV=colorV,
+   #    shareNumLimit=FALSE,
+   #    numLimit=1,
+   #    trimRamp=c(4,3));
+
+   # assign to mem
    mem$geneHitList <- geneHitList;
+   mem$geneHitIM <- geneHitIM;
    mem$geneIM <- geneIM;
    mem$geneIMcolors <- geneIMcolors;
+   mem$geneIMdirection <- geneIMdirection;
 
    #####################################################################
    ## enrichIM incidence matrix using -log10(P-value)